Move BlazeTestSuite to testkit module
