+++ core/src/test/scala/org/http4s/blaze/pipeline/stages/ByteToObjectStageSuite.scala
@@ -18,8 +18,8 @@ package org.http4s.blaze.pipeline.stages
 
 import java.nio.ByteBuffer
 
-import org.http4s.blaze.BlazeTestSuite
 import org.http4s.blaze.pipeline.LeafBuilder
+import org.http4s.blaze.testkit.BlazeTestSuite
 import org.http4s.blaze.util.ImmutableArray
 
 class ByteToObjectStageSuite extends BlazeTestSuite {
