+++ core/src/test/scala/org/http4s/blaze/pipeline/stages/OneMessageStageSuite.scala
@@ -16,8 +16,8 @@
 
 package org.http4s.blaze.pipeline.stages
 
-import org.http4s.blaze.BlazeTestSuite
 import org.http4s.blaze.pipeline.{HeadStage, LeafBuilder, TailStage}
+import org.http4s.blaze.testkit.BlazeTestSuite
 
 import scala.concurrent.Future
